[Bug #19383] win32: Encode timezone name to the locale encoding
On some Windows localizations, the console code page and the active
code page differ.  While Ruby uses the former as the locale encoding,
`_tzset()` in msvcrt seems using the latter internally.  To avoid the
encoding mismatch, achieve the timezone name as a wide char string and
then convert it to the locale encoding.